# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gaia API",
    "description": "Gaia Large Knowledge Model API - manages knowledge graph operations",
    "version": "3.0.0",
    "contact": {
      "name": "Gaia API Support",
      "url": "https://github.com/SiliconEinstein/Gaia/issues"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {"url": "http://localhost:8000", "description": "Local development server"}
  ],
  "paths": {
...
    "/nodes": {
      "get": {
        "summary": "List nodes (paginated)",
        "description": "Get paginated list of nodes with optional type filtering",
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "size", "in": "query", "schema": {"type": "integer", "default": 50}},
          {"name": "type", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {"200": {"description": "List of nodes"}},
        "tags": ["read"]
      }
    },
...
    "/hyperedges": {
      "get": {
        "summary": "List hyperedges (paginated)",
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "size", "in": "query", "schema": {"type": "integer", "default": 50}}
        ],
        "responses": {"200": {"description": "List of hyperedges"}, "503": {"description": "Graph store not available"}},
        "tags": ["read"]
      }
    },
...
    "/search/text": {
      "post": {
        "summary": "Text search (BM25)",
        "description": "Pure BM25 text search without requiring embeddings",
        "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"query": {"type": "string"}, "k": {"type": "integer", "default": 50}}}}}, "required": true},
        "responses": {"200": {"description": "Search results"}},
        "tags": ["search"]
      }
    },
    "/search/nodes": {
      "post": {
        "summary": "Semantic node search",
        "description": "Search nodes using semantic similarity (embedding parameter is deprecated)",
        "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"query": {"type": "string"}, "embedding": {"type": "array", "items": {"type": "number"}, "deprecated": true}, "k": {"type": "integer", "default": 50}}}}}, "required": true},
        "responses": {"200": {"description": "Search results"}},
        "tags": ["search"]
      }
    },
    "/search/hyperedges": {
      "post": {
        "summary": "Semantic hyperedge search",
        "description": "Search hyperedges using semantic similarity (embedding parameter is deprecated)",
        "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"query": {"type": "string"}, "embedding": {"type": "array", "items": {"type": "number"}, "deprecated": true}, "k": {"type": "integer", "default": 50}}}}}, "required": true},
        "responses": {"200": {"description": "Search results"}},
        "tags": ["search"]
      }
    },
...
  },
...
}